separate general dialog, move general dialog to general components, extract action dialog
