+++ src/components/Dialogs/ActionDialog/index.ts
@@ -1,0 +1,3 @@
+import ActionDialog from './ActionDialog'
+
+export default ActionDialog
