+++ src/components/Dialogs/GeneralDialog/index.ts
@@ -1,0 +1,3 @@
+import GeneralDialog from './GeneralDialog'
+
+export default GeneralDialog
