--- src/shared/components/Dialog/index.ts
@@ -1,3 +1,0 @@
-import Dialog from './Dialog'
-
-export default Dialog
