+++ src/shared/components/index.ts
@@ -20,4 +20,3 @@ export { default as CategoryPicker } from './CategoryPicker'
 export { default as Grid, breakpointsOfGrid } from './Grid'
 export { default as HeaderTextField } from './HeaderTextField'
 export { default as Tooltip } from './Tooltip'
-export { default as Dialog } from './Dialog'
